==> scripts/1_Basin_country_overlap.Rmd <==
---
title: "Untitled"
author: "Caitie"
date: "28/10/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(here)
library(sf)
library(raster)
library(countrycode)
library(parallel)
library(pbmcapply)

t_crs<-'ESRI:54009'
```

# Load data

## Countries
```{r message = FALSE}
#gadm<-st_read(here("output_data/GADM36/gadm36_country.shp")) %>% 
#  st_transform(., crs = t_crs)

#gadm<- gadm %>% 
#  st_make_valid(.) %>% 
#  group_by(GID_0) %>% 
#  summarise()

#gadm_area<- gadm %>% 
#  mutate(area_km2 = as.numeric(st_area(.) * 1e-6),
#         Country = countrycode(GID_0, "iso3c", "country.name")) %>% 
#  rename(ISO3 = GID_0) %>% 
#  dplyr::select(ISO3, Country, area_km2, geometry)

#FIX: Some values were not matched unambiguously: XAD, XCA, XCL, XKO, XNC, XPI, XSP

#st_write(gadm_area, here("output_data/GADM36/gadm36_dissolve_area.shp"))

gadm_area<-st_read(here("output_data/GADM36/gadm36_dissolve_area.shp")) %>% 
  st_make_valid() %>% 
  st_transform(., t_crs)

```

## Basins

```{r}
basins_area<-st_read(here("raw_data/Basins/Basins_lv8_mainbasin_moll.shp")) %>% 
  st_make_valid() %>% 
  mutate(b_area_km2 = as.numeric(st_area(.))*1e-6)
```

# Intersect countries and basins

```{r}
continents<-read.csv(here("raw_data/Countries_by_continent.csv"))

cont<-as.list(unique(continents$Continent))
```


```{r message = FALSE}
iso3_basin<-function(x){
  sub_iso3<-continents %>% 
    filter(Continent == x)
  
  sub_gadm<-gadm_area %>% 
  filter(ISO3 %in% sub_iso3$ISO3)

  gadm_basin<-st_intersection(sub_gadm, basins_area)

  gadm_basin_area<-gadm_basin %>% 
    mutate(iso3_b_km2 = as.numeric(st_area(.))*1e-6,
         prop_basin = iso3_b_km2/b_area_km2, #Proportion of basin within each country
         Continent = as.character(x))
  
    
  gadm_basin_df<- gadm_basin_area %>% 
    as.data.frame() %>% 
    dplyr::select(-geometry) 
  
  test<-gadm_basin_df %>% 
    filter(MAIN_BAS == 1080020040)

  n_iso3_basin<- gadm_basin_area %>% 
    as.data.frame() %>% 
    dplyr::select(-geometry) %>% 
    count(MAIN_BAS) #number of countries in each basin
  
  all<-full_join(gadm_basin_df, n_iso3_basin)
  
  gadm_basin_area2<- st_collection_extract(gadm_basin_area, "POLYGON")
  
  st_write(gadm_basin_area2, here(paste0("output_data/Prop_basin_iso3_",x,".shp")), delete_layer = TRUE)
  
  write.csv(all, here(paste0("output_data/Prop_basin_iso3_",x,".csv")))
}

```

```{r}
pbmclapply(cont, iso3_basin, mc.cores = 7, mc.style = "ETA")
```
