==> scripts/4_Trans_land_sediments.Rmd <==
---
title: "Trans_basin_subset"
author: "Caitie"
date: "10/11/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(tidyverse)
library(sf)
library(here)
library(pbmcapply)

options(scipen = 999)

t_crs<-'ESRI:54009'
```

# Description

This code subsets the countries with transboundary basins, match this with sediment export data and pourpoint data to be able to calculate transboundary terrestrial sediment export

# Read in data

Proportin of basin in each country
```{r}
files<-list.files(path = here("output_data"), pattern = "Prop_basin_iso3_", full.names = T)

sub_files<-files[grep(pattern = ".csv", files)]
sub_files<-sub_files[-2] #Remove Antarctica

basin_iso3_files<-files[grep(pattern = ".shp", files)]

shp_files<-list.files(path = here("output_data/Sediment"), pattern = ".shp", full.names = T)
```

# Pourpoint data

```{r}
# All pourpoints
pp_list<-list.files(here("output_data/Pourpoints/"), pattern = "Pourpoints", full.names = T)

sub_pp_list<-pp_list[grep(pp_list, pattern = ".shp")]

pp<-
  do.call(rbind,
          lapply(sub_pp_list, st_read)) %>% 
  st_transform(., t_crs)
```

# Pourpoints in each MAIN BAS

```{r}
basins<-st_read(here("raw_data/Basins/Basins_lv8_mainbasin_moll.shp")) %>% 
  st_make_valid() 

pp_basins<-st_join(pp, basins) %>% 
  filter(!is.na(MAIN_BAS) == T)
```
 
```{r}

sed_dat<-do.call(rbind,
          lapply(shp_files, st_read)) %>% 
  st_transform(., t_crs)
  
  
  #st_read(sub_shp) %>% 
  #dplyr::select(ISO3, Country, MAIN_BAS, Continent, sed_sum, sed_mean) %>% 
  #  st_transform(., t_crs)
```

```{r}
pp_loc<-st_join(pp, sed_dat) 
```

```{r}
# Remove pourpoints without country info
   
   pp_loc_done<-pp_loc %>% 
     filter(!is.na(ISO3) == T) %>% 
     as.data.frame() %>% 
     dplyr::select(-geometry, -sed_mean)

write.csv(pp_loc_done, here("output_data/Pourpoints/PP_location_all.csv"))
   
   # Identify pourpoints that did not match up spatially
   pp_mis<-pp %>% 
     filter(!HYRIV_I %in% pp_loc_done$HYRIV_I)
```

```{r}
n<-10
pp_chunk<-split(pp_mis, factor(sort(rank(row.names(pp_mis))%%n)))

mis_df<-c()

missing_pp<-function(x){ 
    
  pp_mis2<-x
  name<-pp_mis2$HYRIV_I[1]

for(i in 1:nrow(pp_mis2)){

  pp1<-pp_mis2[i,] #shouldn't have any basin/country info
     
     nearest<-st_nearest_feature(pp1, sed_dat)
  
      pp_mis_sed<-sed_dat[nearest,] %>% 
        as.data.frame() %>% 
        dplyr::select(-geometry)
      
      pp_loc_mis<-cbind(pp1, pp_mis_sed) %>% 
     as.data.frame() %>% 
     dplyr::select(-geometry, -sed_mean)
      
      #pp_loc_mis
  # HYRIV_I NEXT_DO  MAIN_RI rv_d_3_ ENDORHE ISO3 Country   MAIN_BAS Continent  sed_sum
#1 10000006       0 10000006   0.065       0  TUN Tunisia 1080031230    Africa 262170.7
      
      test<-left_join(pp_loc_mis, pp_basins, by = c("HYRIV_I", "NEXT_DO", "MAIN_RI", "rv_d_3_", "ENDORHE")) 
      
      #HYRIV_I NEXT_DO  MAIN_RI rv_d_3_ ENDORHE ISO3 Country MAIN_BAS.x Continent  sed_sum
#1 10000006       0 10000006   0.065       0  TUN Tunisia 1080031230    Africa 262170.7
#  MAIN_BAS.y                 geometry
#1 1080031230 POINT (876970.3 4479523)
      
      test<-test %>% 
        mutate(MAIN_BAS.y = ifelse(is.na(MAIN_BAS.y) == T, test$MAIN_BAS.x, test$MAIN_BAS.y))
      
      if(test$MAIN_BAS.x != test$MAIN_BAS.y){
        
        sed_val<-sed_dat %>% 
          filter(MAIN_BAS == test$MAIN_BAS.y,
            ISO3 == test$ISO3) %>% 
          as.data.frame() %>% 
          dplyr::select(sed_sum) %>% 
          as.numeric() 
        
        
        pp_loc_mis<-pp_loc_mis %>% 
          mutate(MAIN_BAS = test$MAIN_BAS.y,
                 sed_sum = sed_val)
        
        
          
        
        # Some don't exist in sed_dat because they are basins on small islands so we remove these
        sub_sed<-sed_dat %>% 
          filter(MAIN_BAS == pp_loc_mis$MAIN_BAS)
        
        if(nrow(sub_sed) == 0){
          pp_loc_mis<-c()
        }
      }
      mis_df<-rbind(mis_df, pp_loc_mis)
 }     
  write.csv(mis_df, here(paste0("output_data/Pourpoints/Missing_pp_chunk", name, ".csv")))
}
```

```{r}
pbmclapply(pp_chunk, missing_pp, mc.cores = 5, mc.style = "ETA")
```

```{r}
done<-rbind(pp_loc_done, mis_df)
   
   # How many rivers still do not have basin/country pourpoint info?
   test<- pp_loc %>% 
     filter(!pp_loc$HYRIV_I %in% done$HYRIV_I)
   
    length(unique(test$HYRIV_I))
   
   # Africa - We still are missing 198 river pourpoint locations or 1.2% of rivers in the dataset
   
   # How many main basins don't have a river pourpoint?
   test2<- sed_dat %>% 
     filter(!sed_dat$MAIN_BAS %in% done$MAIN_BAS) %>% 
     as.data.frame() %>% 
     dplyr::select(-geometry)
   
   length(unique(test2$MAIN_BAS))
   
   test3<-test2 %>% 
     filter(!is.na(sed_sum) == T,
            sed_sum>0)
   
   length(unique(test3$MAIN_BAS))
   
  # Africa - 184 Main basins are in the sediment dataset but still don't have a river pourpoint, 4.7%
  # Africa - 154 of these have some sediment discharge
   
   done2<-done %>% 
     mutate(mis_bas =  length(unique(test2$MAIN_BAS)),
            mis_bas_w_sed = length(unique(test3$MAIN_BAS)),
            mis_pp = length(unique(test$HYRIV_I)),
            total_mis_sed = sum(test3$sed_sum, na.rm = T))
  
   write.csv(done2, here(paste0("output_data/Pourpoints/PP_location_", name2, ".csv")))
   write.csv(test, here(paste0("output_data/Pourpoints/River_location_missing_", name2, ".csv")))
   write.csv(test2, here(paste0("output_data/Pourpoints/Basin_pp_missing_", name2,".csv")))
```

# Determine location of pourpoints

```{r}
continents<-c("Africa", "Asia", "Europe", "North America", "Oceania", "South America")
```

# CHECK BASINS THAT HAVE n = 1 but proportion of basin <1

```{r}
find_pp<-function(x){
   name2<-x
   print(name2)
   
   #match<-case_when(name2 == "Africa" ~ "af",
  #                name2 == "Asia" ~ c("as"),
  #                name2 == "Europe" ~ c("eu", "gr", "si"),
  #                name2 == "North America" ~ "na",
  #                name2 == "Oceania" ~ c("au"),
  #                name2 == "South America" ~ "sa")
 
  #match_pp<-sub_pp_list[grep(paste(match, collapse="|"), sub_pp_list)]
  
  # Read in pourpoint data
 
 #pp_cont<-
#  do.call(rbind,
#          lapply(match_pp, st_read)) %>% 
#   st_transform(., t_crs)
 
 # Match to sediment calculation shapefile for country
  sub_shp<-shp_files[grep(pattern = name2, shp_files)]

  # Read in sediment file
  sed_dat<-st_read(sub_shp) %>% 
  dplyr::select(ISO3, Country, MAIN_BAS, Continent, sed_sum, sed_mean) %>% 
    st_transform(., t_crs)
   
  # Join pourpoint file to sediment file by spatial location
   pp_loc<-st_join(pp_cont, sed_dat) 
   
   # Remove pourpoints without country info
   
   pp_loc_done<-pp_loc %>% 
     filter(!is.na(ISO3) == T) %>% 
     as.data.frame() %>% 
     dplyr::select(-geometry, -sed_mean)
   
   # Identify pourpoints that did not match up spatially
   pp_mis<-pp_cont %>% 
     filter(!HYRIV_I %in% pp_loc_done$HYRIV_I)
   
   # Africa - 3059 rivers don't have a location
   
   ########riv_test<-pp_mis %>% filter(HYRIV_I == 11137695)
   
   # Gap fill pourpoints missing location
   
   mis_df<-c()
   
   # AFRICA - TRY ALL OF THESE TO CORRECT SPECIAL CASES
   
  # i = 238 MAIN_BAS 1080034860 - small island so remove
   # i =  452 MAIN_BAS 1080034680 - small island so remove
   # i = 475 MAIN_BAS 1080034830 - small island so remove
   # i =  489 MAIN_BAS 1080034610 - small island so remove
   # i =  508 MAIN_BAS 1080034760 - small island so remove
   # i =  576 MAIN_BAS 1080034810 - small island so remove
   # i =  584 MAIN_BAS 1080034700 - small island so remove
   # i =  628 MAIN_BAS 1080034820 - small island so remove
   # i =  926 MAIN_BAS 1080026810 - chose wrong basin bc in fjord
   # i =  963 MAIN_BAS 1080026800 - chose wrong basin bc in fjord
   # i = 965 MAIN_BAS 1080026830 - chose wrong basin bc in fjord
   # i =  981 MAIN_BAS 1080034900 - small island so remove
   # i =  1294 MAIN_BAS 1080005420 - chose wrong bc mismatch overlap
   # i =  1769 MAIN_BAS 1080021500 - chose wrong bc mismatch overlap
   # i =  2569 MAIN_BAS 1080011100 - chose wrong bc in fjord
   # i = 2025 MAIN_BAS 1080020040 - chose wrong bc in fjord - this one MAY STILL BE WRONG - selects Angola but maybe should be Congo - is the mouth of the Congo River
   #i = MAIN_BAS 1080020200
 
   for(i in 1:nrow(pp_mis)){
     print(i)
     pp1<-pp_mis[i,]
     
     nearest<-st_nearest_feature(pp1, sed_dat)
  
      pp_mis_sed<-sed_dat[nearest,] %>% 
        as.data.frame() %>% 
        dplyr::select(-geometry)
      
      pp_loc_mis<-cbind(pp1, pp_mis_sed) %>% 
     as.data.frame() %>% 
     dplyr::select(-geometry, -sed_mean)
      
      test<-left_join(pp_loc_mis, pp_basins, by = c("HYRIV_I", "NEXT_DO", "MAIN_RI", "rv_d_3_", "ENDORHE")) 
      
      test<-test %>% 
        mutate(MAIN_BAS.y = ifelse(is.na(MAIN_BAS.y) == T, test$MAIN_BAS.x, test$MAIN_BAS.y))
      
      if(test$MAIN_BAS.x != test$MAIN_BAS.y){
        
        sed_val<-sed_dat %>% 
          filter(MAIN_BAS == test$MAIN_BAS.y,
            ISO3 == test$ISO3) %>% 
          as.data.frame() %>% 
          dplyr::select(-geometry, -ISO3, -Country, -MAIN_BAS, -Continent, -sed_mean) %>% 
          as.numeric() 
        
        
        pp_loc_mis<-pp_loc_mis %>% 
          mutate(MAIN_BAS = test$MAIN_BAS.y,
                 sed_sum = sed_val)
        
        
          
        
        # Some don't exist in sed_dat because they are basins on small islands so we remove these
        sub_sed<-sed_dat %>% 
          filter(MAIN_BAS == pp_loc_mis$MAIN_BAS)
        
        if(nrow(sub_sed) == 0){
          pp_loc_mis<-c()
        }
      }
      
      mis_df<-rbind(mis_df, pp_loc_mis)
   }
   
   #i = 20487
   # We lose some PPs because we exclude those on small islands that are not included in the sediment calculations
   
   # We also lose some PPs because of mismatches in the modelled data
   
   done<-rbind(pp_loc_done, mis_df)
   
   # How many rivers still do not have basin/country pourpoint info?
   test<- pp_loc %>% 
     filter(!pp_loc$HYRIV_I %in% done$HYRIV_I)
   
    length(unique(test$HYRIV_I))
   
   # Africa - We still are missing 198 river pourpoint locations or 1.2% of rivers in the dataset
   
   # How many main basins don't have a river pourpoint?
   test2<- sed_dat %>% 
     filter(!sed_dat$MAIN_BAS %in% done$MAIN_BAS) %>% 
     as.data.frame() %>% 
     dplyr::select(-geometry)
   
   length(unique(test2$MAIN_BAS))
   
   test3<-test2 %>% 
     filter(!is.na(sed_sum) == T,
            sed_sum>0)
   
   length(unique(test3$MAIN_BAS))
   
  # Africa - 184 Main basins are in the sediment dataset but still don't have a river pourpoint, 4.7%
  # Africa - 154 of these have some sediment discharge
   
   done2<-done %>% 
     mutate(mis_bas =  length(unique(test2$MAIN_BAS)),
            mis_bas_w_sed = length(unique(test3$MAIN_BAS)),
            mis_pp = length(unique(test$HYRIV_I)),
            total_mis_sed = sum(test3$sed_sum, na.rm = T))
  
   write.csv(done2, here(paste0("output_data/Pourpoints/PP_location_", name2, ".csv")))
   write.csv(test, here(paste0("output_data/Pourpoints/River_location_missing_", name2, ".csv")))
   write.csv(test2, here(paste0("output_data/Pourpoints/Basin_pp_missing_", name2,".csv")))
  }
```

```{r}
pbmclapply(continents, find_pp, mc.cores = length(continents), mc.style = "ETA")
```

```{r}
pp_location<-list.files(here("output_data/Pourpoints/"), pattern = "PP_location", full.names = T)

pp_final<-
  do.call(rbind,
          lapply(pp_location, read.csv)) 

try<-pp_final %>% 
  filter(MAIN_BAS == 2080005720)
```

# Explore missing basins/rivers

```{r}
name2<-continents[[3]]
done<-read.csv(here(paste0("output_data/Pourpoints/PP_location_", name2, ".csv")))
check_pp<-read.csv(here(paste0("output_data/Pourpoints/River_location_missing_", name2, ".csv")))
check_basins<-read.csv(here(paste0("output_data/Pourpoints/Basin_pp_missing_", name2,".csv")))

sub_shp<-shp_files[grep(pattern = name2, shp_files)]

  # Read in sediment file
  sed_dat<-st_read(sub_shp) %>% 
  dplyr::select(ISO3, Country, MAIN_BAS, Continent, sed_sum, sed_mean) %>% 
    st_transform(., t_crs)
  
  transboundary<-read.csv(here("output_data/Transboundary/Terrestrial_transboundary_Europe.csv"))
  
  prop<-read.csv(here("output_data/Prop_basin_iso3_Europe.csv"))
  prop_europe<-prop %>% 
    filter(Continent == "Europe")
  
  sub_check_basins<-check_basins %>% 
    filter(MAIN_BAS %in% prop_europe$MAIN_BAS)
  
  test<-full_join(check_basins, prop_europe)
  
  
```

```{r}
length(unique(check_basins$MAIN_BAS)) # Number of basins without pourpoint
length(unique(check_pp$HYRIV_I)) # Number of rivers without pourpoint location
length(unique(sed_dat$MAIN_BAS)) #Number of basins in sediment data

length(unique(check_basins$MAIN_BAS))/length(unique(sed_dat$MAIN_BAS))

some_sed<-check_basins %>% 
  filter(sed_sum>0)

length(unique(some_sed$MAIN_BAS))
length(unique(some_sed$MAIN_BAS))/length(unique(sed_dat$MAIN_BAS))
sum(some_sed$sed_sum, na.rm = T)
```
Europe 36% of basins are missing pourpoints, 30% that have sediment run off, missing 1519623331 of sediment

Africa

South America

Asia

Oceania

North America
# Calculate mismatches in sediment and pourpoint locations

```{r}
pp_loc_files<-list.files(here("output_data/Pourpoints"), pattern = "PP", full.names = T)
sub_pp_loc_files<-pp_loc_files[grep(pattern = ".csv", pp_loc_files)]
```

```{r}
# Testing the data to see which basins still don't have pourpoints
  name2<-x
  match_sed<-shp_files[grep(pattern = name2, shp_files)]

  sed_dat<-st_read(match_sed) %>% 
  dplyr::select(ISO3, Country, MAIN_BAS, Continent, sed_sum, sed_mean) %>% 
    st_transform(., t_crs)
  
  pp_match<-sub_pp_loc_files[grep(pattern = name2, sub_pp_loc_files)]
  
  pp_dat<-read.csv(pp_match) 
  
  # Calculate total sediment by main basin
  total_sed<-sed_dat %>% 
  as.data.frame() %>% 
  dplyr::select(-geometry) %>% 
  group_by(MAIN_BAS) %>% 
  summarise(total_sed = sum(sed_sum, na.rm = T))
  
  
  # Join total sediment to sediment data to calculate proportion of sediment per country/basin
  sed_basin<-sed_dat %>% 
  full_join(., total_sed, by = "MAIN_BAS") %>% 
  mutate(prop_sed = sed_sum/total_sed)
  
    # Check to make sure proportions summing correctly
  total_prop<-sed_basin %>% 
  as.data.frame() %>% 
  dplyr::select(-geometry) %>% 
  group_by(MAIN_BAS) %>% 
  summarise(total_prop = sum(prop_sed, na.rm = T))
  
  sed_basin_loc<-full_join(sed_basin, pp_dat, by = c("MAIN_BAS", "Continent"))
  
  #Check to make sure all basins have pp location data
  length(which(is.na(sed_basin_loc$HYRIV_I) == T))
  
  
  # Filter pp with no location
  test<-sed_basin_loc %>% filter(is.na(HYRIV_I) == T)
  
  length(unique(test$MAIN_BAS))
  
  #185 missing location data
  
  # Filter pp's with locations
  test2<-sed_basin_loc %>% filter(!is.na(HYRIV_I) == T)
  
  # See how many pps without location in that country have a location in another country
  
  test4<- test %>% filter(!MAIN_BAS %in% test2$MAIN_BAS)
  
  length(unique(test4$MAIN_BAS))
  
  test5<-test4 %>% 
    filter(!is.na(total_sed) == T,
           total_sed > 0)
  
  length(unique(test5$MAIN_BAS))
  
  xy<-test4 %>% 
    filter(is.na(total_sed) == T | total_sed <= 0)
    
  
  #154 of these have some level of sedimentation
  
 dat<-read.csv("/home/kuempel/transboundary_sediment/output_data/Prop_basin_iso3_Africa.csv")
 
 test6<-left_join(test5, dat, by= c("Country.x" = "Country", "ISO3.x" = "ISO3", "MAIN_BAS", "Continent"))
 
 # Need to figure out why these basins don't have pourpoints and if this is a problem - the ones I spot checked seem to be small coastal basins.
 
 # Transboundary basins without pourpoints
 # There is just 1 now.
 test7<-test6 %>% filter(n>1)
 
 length(unique(test7$MAIN_BAS))
 
 # Join pp_dat to basins/pp to see where mismatches occur
 
 test8<-full_join(pp_dat, pp_basins, by = c("HYRIV_I", "NEXT_DO", "MAIN_RI", "rv_d_3_", "ENDORHE"))
 
 test9 <- test8 %>% filter(MAIN_BAS.x != MAIN_BAS.y) 
 
 
 
 
```


MAIN BAS
1080001370 - edge doesn't line up - should be EGYPT
1080001390 - edge doesn't line up - should be SUDAN
1080005080 - edge doesn't line up - should be SOMALIA
1080005140 - edge doesn't line up - should be SOMALIA
1080005300 
1080008270 
1080008880 
1080012210
1080012680 
1080013420 
1080013640 
1080018110 
1080020040 
1080020200 
1080020570 
1080021500
1080023020 
1080023100 
1080023180 
1080023190 
1080026800 
1080026810 
1080026830 
1080027980
1080028850 
1080028870 
1080028900 
1080030250 
1080031100 
1080031570
