==> scripts/5_Sediment_at_pourpoints.Rmd <==
---
title: "Untitled"
author: "Caitie"
date: "02/02/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(here)
library(tidyverse)
library(sf)
library(raster)

t_crs<-'ESRI:54009'
```
# Read in basin/pourpoint data

Pourpoints

```{r}
# All pourpoints
pp_list<-list.files(here("output_data/Pourpoints/"), pattern = "Pourpoints", full.names = T)

sub_pp_list<-pp_list[grep(pp_list, pattern = ".shp")]

pp<-
  do.call(rbind,
          lapply(sub_pp_list, st_read)) %>% 
  st_transform(., t_crs)
```

```{r}
list_sed_basin<-list.files(here("output_data/Sediment"), pattern = ".shp", full.names = T)

#list_sed_basin<-list.files(here("output_data/Transboundary"), pattern = #"Terrestrial_transboundary_", full.names = TRUE)

sed_dat <- 
  do.call(rbind,
          lapply(list_sed_basin, st_read)) %>% 
  as.data.frame() %>% 
  dplyr::select(-geometry)

bas_sed_dat<-sed_dat %>% 
  group_by(MAIN_BAS) %>% 
  summarise(bas_sed_sum = sum(sed_sum, na.rm = T))
```

# PP Location

```{r}
pp_loc_list<-list.files(here("output_data/Pourpoints"), pattern = "Missing_pp_chunk", full.names = T)

pp_loc_mis<-
  do.call(rbind,
          lapply(pp_loc_list, read.csv))

pp_all<-read.csv(here("output_data/Pourpoints/PP_location_all.csv"))

pp_done<-rbind(pp_all, pp_loc_mis)
```

# Sediment distribution

```{r}
#dist_sed<-function(x){
#name2<-x
#dat<-read.csv(here(paste0("output_data/Pourpoints/PP_location_", name2, ".csv"))) %>% 
#  dplyr::select(-sed_sum, -mis_bas, -mis_bas_w_sed, -mis_pp, -total_mis_sed)

# Join basin and pourpoint info

bas_pp<-full_join(pp, pp_done) %>% #use to be dat
  filter(!is.na(MAIN_BAS) == T) %>% 
  left_join(., bas_sed_dat)

# Distribute sediments

total_dis<-bas_pp %>% 
  as.data.frame() %>% 
  dplyr::select(-geometry) %>% 
  group_by(MAIN_BAS) %>% 
  summarise(Total_discharge = sum(rv_d_3_, na.rm = T))  %>% 
  as.data.frame()
  
bas_pp_dist<- bas_pp %>% 
  left_join(., total_dis, by = "MAIN_BAS") %>% 
  mutate(prop_dis = rv_d_3_/Total_discharge,
         sed_pp = prop_dis * bas_sed_sum) %>% 
  dplyr::select(-X)


#test<-bas_pp_dist %>% 
#  filter(MAIN_BAS == 1080031170)

st_write(bas_pp_dist,here(paste0("output_data/Sediment_at_pourpoint/Sediment_at_pourpoint.gpkg")), driver = "GPKG", delete_layer = T)

save(data1, file = "data.RData")
```

For loop to save

```{r}
continents<-c("Africa", "Asia", "Europe", "North America", "Oceania", "South America")
```


```{r}
for(i in 1:length(continents)){
  
  name<- continents[i]
  
  sub<-bas_pp_dist %>% 
    filter(Continent == name)
  
  st_write(sub,here(paste0("output_data/Sediment_at_pourpoint/Sediment_at_pourpoint_",name, ".shp")), delete_layer = T)
  
  st_write(sub,here(paste0("output_data/Sediment_at_pourpoint/Sediment_at_pourpoint_",name, ".shp")), driver = "GeoJSON", delete_layer = T)
  
}
```

```{r}
#pbmclapply(continents, dist_sed, mc.cores = length(continents), mc.style = #"ETA")
```


```{r}
test<-list.files(path = here("output_data/Sediment_at_pourpoint/"), pattern = "Sediment_at_pourpoint")
test<-st_read(here(paste0("output_data/Sediment_at_pourpoint/Sediment_at_pourpoint.shp")))

test2<-test[1:10,]
```
