==> scripts/8_Figure1.Rmd <==
---
title: "Untitled"
author: "Caitie"
date: "23/11/2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(here)
library(tidyverse)
library(ggplot2)
library(tmap)
library(sf)
library(patchwork)
library(rnaturalearth)
library(rnaturalearthdata)

t_crs<-'ESRI:54009'

```

# Load data

```{r}
basin_table<-read.csv(here("results/Basin_results_table.csv"))
```


Basin shapefile
```{r}
basins<-st_read(here("raw_data/Basins/Basins_lv8_mainbasin_moll.shp")) %>% 
  st_make_valid()
```
# Country shapefile

```{r}
world <- ne_countries(scale = "medium", returnclass = "sf") %>% 
  dplyr::select(iso_a3) %>% 
  rename(ISO3 = iso_a3) %>% 
  st_transform(., "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs")
```

# Figure 1

```{r}
shp_basins<-full_join(basins, basin_table) %>% 
  #st_transform("+proj=eck4 +lon_0=0 +x_0=0 +y_0=0 +datum=WGS84 +units=m +no_defs") %>% 
  st_transform(., "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs")
```

```{r}
breaks<-c(0,2,4,6,8,10,12,14,16,18,20)

test<-shp_basins %>% 
  filter(is.na(n_iso3) == T)


#Needs to be categorical to have histogram

n_map<-tm_shape(shp_basins) +
    tm_polygons("n_iso3", 
                breaks = breaks,
                style = "cont", 
                title = "         Number of countries",
                palette = "Greens",
            colorNA = NA,
            textNA = NA,
            legend.is.portrait = FALSE) +
  tm_borders(lwd = 0.7,
             alpha = 0.5)+
  #tm_legend(legend.position = c("bottom")) +
  tm_layout(legend.outside = T,
            legend.outside.position = "bottom",
            legend.position = c(0.35, 0.1),
            frame = F) +
  tm_shape(world) +
   tm_borders() 
```

```{r}
n_map
```

Figure 1A
```{r}
tmap_save(n_map, here("figures/Global_countries_per_basin_Figure1A.png"), dpi = 300)
```

# Sediment within transboundary watersheds

```{r}

trans_basins<-shp_basins %>% 
  mutate(total_sed_bas2<-ifelse(n_iso3>1, total_sed_bas, NA),
    total_sed_bas_mt = total_sed_bas2/1000000)


#Needs to be categorical to have histogram

trans_sed_map<-tm_shape(trans_basins) +
    tm_fill("total_sed_bas_mt", 
              #  breaks = breaks,
                style = "cont", 
                title = "         Sediment (Mt)",
                palette = "-RdYlBu",
            colorNA = "lightgrey",
            textNA = "Not transboundary",
            legend.is.portrait = FALSE) +
  tm_borders(lwd = 0.9) +
  #tm_legend(legend.position = c("bottom")) +
  tm_shape(world) +
   tm_borders(col = "black",
              lwd = 0.1,
              alpha = 0.5) +
    tm_layout(legend.outside = T,
            legend.outside.position = "bottom",
            legend.position = c(0.1, 0.1),
            frame = F) 
```

```{r}
trans_sed_map
```

Figure 1B
```{r}
tmap_save(trans_sed_map, here("figures/Global_sediment_per_trans_basin.png"), dpi = 300)
```

# Country terrestrial transboundary sediment

```{r}
terr_files<-list.files(here("output_data/Sediment/"), pattern = ".shp", full.names = TRUE)

terr_sed<-do.call(rbind,
                  lapply(terr_files, st_read))
```

```{r}
trans_basins<-shp_basins %>% 
  filter(n_iso3 >1)

trans_terr_sed<-terr_sed %>% 
  mutate(sed_sum2 = ifelse(MAIN_BAS %in% trans_basins$MAIN_BAS, sed_sum/1000000, NA)) %>% 
  st_transform(., "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs") %>% 
  dplyr::select(ISO3, Country, MAIN_BAS, sed_sum, sed_sum2)
```

```{r}
terr_sed_map<-tm_shape(trans_terr_sed) +
    tm_fill("sed_sum2", 
               # breaks = breaks,
                style = "cont", 
                title = "        Sediment (Mt)",
                palette = "-RdYlBu",
            colorNA = "lightgrey",
            textNA = "Not transboundary",
            legend.is.portrait = FALSE) +
  #tm_borders(lwd = 0.75)+
  #tm_legend(legend.position = c("bottom")) +
  tm_layout(legend.outside = T,
            legend.outside.position = "bottom",
            legend.position = c(0.1, 0.1),
            frame = F) +
  tm_shape(shp_basins) +
    tm_borders()#+
 # tm_shape(world) +
  #  tm_borders(lwd = 0.25)
  

```



```{r}
terr_sed_map
```

Figure not used in final version
```{r}
#tmap_save(terr_sed_map, here("figures/Global_sediment_per_trans_iso3_basin.png"), dpi = 300)
```



# EEZ sediment transport

```{r}
eez_mismatch<-read.csv(here("output_data/EEZ_sediments/EEZ_sed_trans_mismatch.csv"))
```

```{r}
sink<-eez_mismatch %>% 
  group_by(Sink_ISO3) %>% 
  summarise(total_sed = sum(total_eez_sed, na.rm = T))
```

```{r}
eez<-st_read(here("raw_data/EEZ/World_EEZ_v11_20191118/eez_v11.shp")) %>% 
  st_make_valid(., NA_on_exception = T) %>% 
  st_transform(., t_crs)

eez2<-eez %>% 
  mutate(area_m2 = as.numeric(st_area(.))) %>% 
  group_by(ISO_TER1) %>% 
  summarise(total_area = sum(area_m2, na.rm = T))
```


```{r}
eez<-st_read(here("raw_data/EEZ/eez_simple.shp")) %>% 
  st_transform(., "+proj=cea +lon_0=0 +lat_ts=30 +x_0=0 +y_0=0 +datum=WGS84 +ellps=WGS84 +units=m +no_defs")
```

```{r}
eez_sink<-full_join(eez, sink, by = c("rgn_key" = "Sink_ISO3"))
```

Missing EEZs
ABW - Aruba
AZE - Azerbeijan
BLM
CUW
GLP
MTQ
PRI
PSE
TKM
```{r}

test<-sub_eez_sink %>% 
  filter(total_sed_mil>60)

breaks_eez<-c(0,0.00000001,0.000001,0.001,0.1,1, 10, 25, 50, 75, 100, 150)
breaks_eez<-as.numeric(quantile(sub_eez_sink$total_sed_mil, na.rm = T,probs = seq(0, 1, 0.1)))

sub_eez_sink<-eez_sink %>% 
  filter(!is.na(rgn_name) == T) %>% 
  mutate(total_sed_mil = total_sed/1000000)#,
       #  total_sed_mil = ifelse(is.na(total_sed_mil) == T, 0, total_sed_mil)) #%>% 
 # mutate(total_sed = ifelse(is.na(total_sed) == T, 0, total_sed))

eez_map<-tm_shape(sub_eez_sink) +
    tm_polygons("total_sed_mil", 
                breaks = breaks_eez,
                style = "cont",
                palette = "-RdYlBu",
            textNA = NA,
           # style = "log10",
             title = "                                                         Sediment (Million Mt)",
            legend.is.portrait = FALSE) +
  tm_borders(lwd = 1)+
  tm_layout(legend.outside = T,
            legend.outside.position = "bottom",
            legend.position = c(0.1,0.1),
            frame = F) 

eez_map
```

Figure 1C
```{r}
eez_map2<-eez_map +
  tm_shape(world) +
   tm_borders() 

eez_map2

```
```{r}
tmap_save(eez_map, here("figures/EEZ_sediment_totals_Figure1C.png"), dpi = 300)
```

```{r}
#combined<-n_map + eez_map
```


```{r}
#tmap_save(combined, here("figures/EEZ_sediment_Country_basin_combined.png"), dpi = 300)

```
